Move providerconfitypes into their own package
